requirements.txt
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Login
gunicorn